Calendar: Split view model files into separate directory
calendar-model/ was getting crowded.
