+++ src/components/Line.ts
@@ -1,4 +1,4 @@
-import { RenderedLineInfo } from '../calendar-model/scroll-view-model'
+import { RenderedLineInfo } from '../view-model/scroll-view-model'
 import displayRange from '../display-range'
 import textFilter from '../text-filter'
