+++ src/view-model/scroll-view-model.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava'
-import { LeiningGenerator } from './generator.ts'
-import { UserSettings } from './user-settings.ts'
+import { LeiningGenerator } from '../calendar-model/generator.ts'
+import { UserSettings } from '../calendar-model/user-settings.ts'
 import { RenderedEntry, ScrollViewModel } from './scroll-view-model.ts'
 import { renderLine } from './test-utils.ts'
